Move the ${input:…} prompt into taskui.ts — last of group 2
It goes into an existing module rather than a new one, because it is the last
step of the same launch the picker and the tasks panel start, and they are its
only callers. PLAN's rule for shared state — move it to the module that owns
it, not to a new home by reflex — reads the same way for behaviour.

Moved whole (two guarded ranges): inputCtx, openInputPrompt,
closeInputPrompt, submitInputPrompt and the three $("in*") listeners.

The point of the move is that it needed no hook and *deleted* two. It was a
member of taskui's host object and of palui's, both because it lived in
main.ts; now taskui calls it directly and palui imports it. taskui's host is
down to six members, palui's to eleven. That is seam rule 1 — move the callee
down and the machinery disappears — applied to something already extracted.

Verified mechanical: the relocated block diffed against
`git show HEAD:src/main.ts` differs only by two `export` keywords.
main.ts 1,724 -> 1,682. Gates: tsc --noEmit clean, 368 vitest pass. No Rust
touched.
